BIG-FIX AND UPDATE: fix Validation user , account , project, subdivision, villa

+++ src/controllers/accounts.controller.ts
@@ -1,12 +1,13 @@
 import { Request, Response } from 'express';
 import { ParamsDictionary } from 'express-serve-static-core';
+import { RoleName } from '~/constants/enum';
 import { USERS_MESSAGES } from '~/constants/message';
 import { createAccountReq, updateAccountReq } from '~/models/requests/account.request';
 import { TokenPayload } from '~/models/requests/register.request';
 import usersService from '~/services/users.services';
 export const createAccountController = async (req: Request<ParamsDictionary, any, createAccountReq>, res: Response) => {
   const { user_id } = req.decoded_authorization as TokenPayload;
-  if ((await usersService.getRole(user_id)) !== 'admin') {
+  if ((await usersService.getRole(user_id)) !== RoleName.ADMIN) {
     throw new Error(USERS_MESSAGES.USER_NOT_ACCESS);
   }
   const result = await usersService.createAccount(req.body);
@@ -18,6 +19,10 @@ export const createAccountController = async (req: Request<ParamsDictionary, any
 //update account dành cho admin by id
 export const updateAccountController = async (req: Request<ParamsDictionary, any, updateAccountReq>, res: Response) => {
   const { id } = req.params;
+  const { user_id } = req.decoded_authorization as TokenPayload;
+  if ((await usersService.getRole(user_id)) !== RoleName.ADMIN) {
+    throw new Error(USERS_MESSAGES.USER_NOT_ACCESS);
+  }
   const result = await usersService.updateAccountById(id, req.body);
   return res.json({
     message: USERS_MESSAGES.UPDATE_ACCOUNT_SUCCESS,
@@ -26,6 +31,10 @@ export const updateAccountController = async (req: Request<ParamsDictionary, any
 };
 //get account dành cho admin by id
 export const getAccountController = async (req: Request, res: Response) => {
+  const { user_id } = req.decoded_authorization as TokenPayload;
+  if ((await usersService.getRole(user_id)) !== RoleName.ADMIN) {
+    throw new Error(USERS_MESSAGES.USER_NOT_ACCESS);
+  }
   const result = await usersService.getAccount();
   return res.json({
     message: USERS_MESSAGES.GET_ACCOUNT_SUCCESS,
@@ -34,6 +43,10 @@ export const getAccountController = async (req: Request, res: Response) => {
 };
 export const deleteAccountController = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const { user_id } = req.decoded_authorization as TokenPayload;
+  if ((await usersService.getRole(user_id)) !== RoleName.ADMIN) {
+    throw new Error(USERS_MESSAGES.USER_NOT_ACCESS);
+  }
   const result = await usersService.deleteAccountById(id);
   return res.json({
     message: USERS_MESSAGES.DELETE_ACCOUNT_SUCCESS,
